Create Trees
This C++ program implements a basic stack using a linked list. The Node class has an integer Value and a pointer Next to the next node. The addAtTheFront function adds a new node to the top of the stack. The printStack function prints all values in the stack from top to bottom. The main function initializes the stack with one element, adds another element to the front, and prints the stack.